+++ qom/loopers/__init__.py
@@ -1,3 +1,3 @@
-from qom.ui.log import init_log
-
-init_log()
+from qom.loopers.XLooper import XLooper
+from qom.loopers.XYLooper import XYLooper
+from qom.loopers.XZLooper import XZLooper