+++ qom/systems/DODMSystem.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.systems.DODMSystem'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-12-04'
-__updated__ = '2020-12-04'
+__updated__ = '2020-12-21'
 
 # dependencies
 import logging
@@ -24,17 +24,16 @@ class DODMSystem(BaseSystem):
         
     Parameters
     ----------
-    data : dict
-        Data for the system.
+    params : dict
+        Parameters for the system.
     """
 
-    def __init__(self, data):
+    def __init__(self, params):
         """Class constructor for DODMSystem."""
 
         # initialize super class
-        super().__init__(data)
+        super().__init__(params)
 
-        # initialize properties
-        self.name = data.get('name', 'DODMSystem')
-        self.code = data.get('code', 'dodms')
-        self.params = data.get('params', {})
+        # set attributes
+        self.code = 'dodms'
+        self.name = 'DODMSystem'